# specs/001-network-simulator-platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: NetLab API
  version: 1.0.0
  description: Shared single-host network-simulator control API.
servers:
  - url: /api/v1
security: []
paths:
...
  /links/{linkId}/reconnect:
    post:
      operationId: reconnectLink
      parameters: [{name: linkId, in: path, required: true, schema: {type: string, format: uuid}}, {$ref: '#/components/parameters/IfMatch'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [retained_endpoint_id, replacement_endpoint_id], properties: {retained_endpoint_id: {type: string, format: uuid}, replacement_endpoint_id: {type: string, format: uuid}}}}}}
      responses:
        '202': {$ref: '#/components/responses/AcceptedTask'}
        '412': {$ref: '#/components/responses/PreconditionFailed'}
...
  /captures:
    get:
      operationId: listCaptures
      parameters: [{name: laboratory_id, in: query, schema: {type: string, format: uuid}}, {name: include_internal, in: query, schema: {type: boolean, default: false}}]
      responses:
        '200': {description: Capture sessions, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/CaptureSession'}}}}}
    post:
      operationId: startCapture
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/StartCapture'}}}}
      responses:
        '202': {$ref: '#/components/responses/AcceptedTask'}
  /captures/{captureId}:
    get:
      operationId: getCapture
      parameters: [{name: captureId, in: path, required: true, schema: {type: string, format: uuid}}]
      responses:
        '200': {description: Capture metadata, content: {application/json: {schema: {$ref: '#/components/schemas/CaptureSession'}}}}
    delete:
      operationId: stopCapture
      parameters: [{name: captureId, in: path, required: true, schema: {type: string, format: uuid}}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses:
        '202': {$ref: '#/components/responses/AcceptedTask'}
  /captures/{captureId}/stream:
    get:
      operationId: streamCapture
      parameters: [{name: captureId, in: path, required: true, schema: {type: string, format: uuid}}]
      responses:
        '200': {description: Live packet stream, content: {application/vnd.tcpdump.pcap: {schema: {type: string, contentMediaType: application/vnd.tcpdump.pcap}}}}
  /client-tools/wireshark-helper/{platform}:
    get:
      operationId: downloadWiresharkHelper
      parameters:
        - name: platform
          in: path
          required: true
          schema: {enum: [linux-amd64, windows-amd64, darwin-amd64, darwin-arm64]}
      responses:
        '200':
          description: Platform-specific NetLab Wireshark helper executable
          headers:
            X-Checksum-SHA256: {schema: {type: string}}
          content:
            application/octet-stream: {schema: {type: string, format: binary}}
        '404': {description: Helper package unavailable for the requested platform}
  /traffic-filters:
    get:
      operationId: listTrafficFilters
      parameters: [{name: laboratory_id, in: query, schema: {type: string, format: uuid}}]
      responses:
        '200': {description: Traffic Filter sessions with ambiguity metadata, content: {application/json: {schema: {type: array, items: {type: object}}}}}
    post:
      operationId: startTrafficFilter
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/StartTrafficFilter'}}}}
      responses:
        '202': {$ref: '#/components/responses/AcceptedTask'}
  /traffic-filters/{filterId}:
    get:
      operationId: getTrafficFilter
      parameters: [{name: filterId, in: path, required: true, schema: {type: string, format: uuid}}]
      responses:
        '200': {description: Filter and observed path, content: {application/json: {schema: {type: object}}}}
    delete:
      operationId: stopTrafficFilter
      parameters: [{name: filterId, in: path, required: true, schema: {type: string, format: uuid}}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses:
        '202': {$ref: '#/components/responses/AcceptedTask'}
...
  /artifacts/{artifactId}:
    get:
      operationId: downloadArtifact
      parameters: [{name: artifactId, in: path, required: true, schema: {type: string, format: uuid}}]
      responses:
        '200': {description: Artifact bytes, content: {application/octet-stream: {schema: {type: string, format: binary}}}}
...
components:
  parameters:
...
    IfMatch: {name: If-Match, in: header, required: true, schema: {type: string}, description: Quoted resource revision}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: false, schema: {type: string, minLength: 8, maxLength: 128}}
  responses:
    AcceptedTask: {description: Accepted, content: {application/json: {schema: {$ref: '#/components/schemas/TaskEnvelope'}}}}
...
    PreconditionFailed: {description: Revision mismatch, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
...
  schemas:
...
    Node:
      type: object
      required: [id, laboratory_id, name, kind, revision, desired_state, observed_state]
      properties: {id: {type: string, format: uuid}, laboratory_id: {type: string, format: uuid}, name: {type: string}, kind: {enum: [qemu, docker, pc, switch_l2, switch_l3]}, template_version_id: {type: string, format: uuid}, revision: {type: integer}, desired_state: {enum: [stopped, running, deleted]}, observed_state: {enum: [unknown, provisioning, starting, running, stopping, stopped, failed, deleting]}, cpu_count: {type: integer}, cpu_quota_micros: {type: integer}, memory_mib: {type: integer}, storage_gib: {type: integer}, interface_limit: {type: integer}, process_limit: {type: integer}, config: {type: object}, last_error: {$ref: '#/components/schemas/Problem'}}
...
    NodeInterface:
      type: object
      required: [id, node_id, slot, name, driver, mac_address, revision]
      properties: {id: {type: string, format: uuid}, node_id: {type: string, format: uuid}, slot: {type: integer}, name: {type: string}, driver: {type: string}, mac_address: {type: string}, desired_link_id: {type: string, format: uuid}, operational_state: {type: string}, revision: {type: integer}}
    Link:
      type: object
      required: [id, laboratory_id, endpoint_a_id, endpoint_b_id, revision, desired_state, observed_state]
      properties: {id: {type: string, format: uuid}, laboratory_id: {type: string, format: uuid}, endpoint_a_id: {type: string, format: uuid}, endpoint_b_id: {type: string, format: uuid}, revision: {type: integer}, desired_state: {type: string}, observed_state: {type: string}}
...
    PortMapping:
      type: object
      required: [id, node_id, protocol, host_address, host_port, guest_address, guest_port, revision, observed_state, created_at]
      properties: {id: {type: string, format: uuid}, node_id: {type: string, format: uuid}, protocol: {enum: [tcp, udp]}, host_address: {type: string}, host_port: {type: integer, minimum: 1, maximum: 65535}, guest_address: {type: string}, guest_port: {type: integer, minimum: 1, maximum: 65535}, revision: {type: integer, minimum: 1}, observed_state: {enum: [pending, active, failed, deleting]}, last_error: {$ref: '#/components/schemas/Problem'}, created_at: {type: string, format: date-time}}
...
    StartCapture:
      type: object
      required: [source_type, source_id]
      properties: {laboratory_id: {type: string, format: uuid}, source_type: {enum: [interface, link]}, source_id: {type: string, format: uuid}, interface: {type: string}, filter: {type: string}, format: {enum: [pcap, pcapng]}, retain: {type: boolean, default: false}, duration_seconds: {type: integer, maximum: 900}, max_bytes: {type: integer, maximum: 268435456}}
    CaptureSession:
      type: object
      required: [id, source_type, source_id, format, state, retain, max_bytes, bytes_written, packets, truncated, created_at]
      properties: {id: {type: string, format: uuid}, laboratory_id: {type: string, format: uuid}, source_type: {enum: [interface, link]}, source_id: {type: string, format: uuid}, purpose: {type: string}, parent_resource_id: {type: string, format: uuid}, filter: {type: string}, format: {enum: [pcap, pcapng]}, state: {enum: [requested, starting, streaming, stopping, completed, cancelled, failed, truncated]}, retain: {type: boolean}, max_bytes: {type: integer}, bytes_written: {type: integer}, packets: {type: integer}, truncated: {type: boolean}, artifact_id: {type: string, format: uuid}, artifact_url: {type: string}, expires_at: {type: string, format: date-time}, completion_reason: {type: string}, created_at: {type: string, format: date-time}, started_at: {type: string, format: date-time}, finished_at: {type: string, format: date-time}, last_error: {$ref: '#/components/schemas/Problem'}}
    StartTrafficFilter:
      type: object
      required: [laboratory_id, match]
      properties: {laboratory_id: {type: string, format: uuid}, interface_ids: {type: array, items: {type: string, format: uuid}}, link_ids: {type: array, items: {type: string, format: uuid}}, match: {$ref: '#/components/schemas/TrafficFilterExpression'}, max_observations: {type: integer, minimum: 1, maximum: 100000}, color: {type: string, pattern: '^#[0-9a-fA-F]{6}$', default: '#f59e0b'}}
    TrafficFilterExpression:
      type: object
      description: A leaf predicate or exactly one recursive boolean operator. AND/OR require at least two operands; nesting depth is limited to 32.
      properties:
        source_address: {type: string}
        destination_address: {type: string}
        protocol: {enum: [tcp, udp, icmp, icmp6, arp, ip, ip6]}
        source_port: {type: integer, minimum: 1, maximum: 65535}
        destination_port: {type: integer, minimum: 1, maximum: 65535}
        and: {type: array, minItems: 2, items: {$ref: '#/components/schemas/TrafficFilterExpression'}}
        or: {type: array, minItems: 2, items: {$ref: '#/components/schemas/TrafficFilterExpression'}}
        not: {$ref: '#/components/schemas/TrafficFilterExpression'}
    OperationTask:
      type: object
      required: [id, kind, resource_type, resource_id, state, progress_current, progress_total, created_at]
      properties: {id: {type: string, format: uuid}, kind: {type: string}, resource_type: {type: string}, resource_id: {type: string}, requested_revision: {type: integer}, state: {enum: [queued, running, cancelling, succeeded, failed, cancelled]}, progress_current: {type: integer}, progress_total: {type: integer}, result: {type: object}, error: {$ref: '#/components/schemas/Problem'}, created_at: {type: string, format: date-time}, started_at: {type: string, format: date-time}, finished_at: {type: string, format: date-time}}
...
    TaskEnvelope:
      type: object
      required: [task]
      properties: {task: {$ref: '#/components/schemas/OperationTask'}, node: {$ref: '#/components/schemas/Node'}, link: {$ref: '#/components/schemas/Link'}, interface: {$ref: '#/components/schemas/NodeInterface'}, port_mapping: {$ref: '#/components/schemas/PortMapping'}, artifact: {$ref: '#/components/schemas/Artifact'}}
    Problem:
      type: object
      properties: {code: {type: string}, message: {type: string}, retryable: {type: boolean}, resource_type: {type: string}, resource_id: {type: string}, task_id: {type: string, format: uuid}, phase: {type: string}, cleanup: {type: string}, operator_hint: {type: string}, retry_after_seconds: {type: integer, minimum: 1}, details: {type: object}}